LayoutTests/accessibility/mac/checked-status-tree-items-expected.txt

item
item
This tests that aria-checked is supported on tree items and changes generate notifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
tree item not checked: value attribute supported: false
value attribute supported: true
initial tree item value: 1
final tree item value: 0
PASS successfullyParsed is true

TEST COMPLETE
